requirements.txt
# FESTA Evaluation Requirements

# Core ML/DL
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
accelerate>=0.24.0
bitsandbytes>=0.41.0

# Vision
Pillow>=10.0.0
opencv-python>=4.8.0

# Data & ML
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Datasets
datasets>=2.14.0

# Metrics & Evaluation
scikit-image>=0.21.0

# Utilities
tqdm>=4.66.0
PyYAML>=6.0
python-dotenv>=1.0.0

# API Integration (for complement generation)
openai>=1.0.0
google-generativeai>=0.3.0

# Optional (for advanced features)
# psutil>=5.9.0
# pynvml>=11.5.0
# PyPDF2>=3.0.0

